refactor(frontend): modularize legacy app orchestration
