+++ src/components/providers/forms/GrokBuildProviderForm.tsx
@@ -28,7 +28,10 @@ import type {
   ProviderCategory,
   ProviderMeta,
 } from "@/types";
-import type { ProviderFormProps, ProviderFormValues } from "./ProviderForm";
+import type {
+  ProviderFormProps,
+  ProviderFormValues,
+} from "./ProviderForm.types";
 import { BasicFormFields } from "./BasicFormFields";
 import { CodexFormFields } from "./CodexFormFields";
 import { ProviderPresetSelector } from "./ProviderPresetSelector";
